+++ src/lib.rs
@@ -27,7 +27,7 @@ use crate::indicators::r038::R038;
 use crate::indicators::r048::R048;
 use crate::indicators::r058::R058;
 use crate::indicators::util::{SecondLowestBidRatio, Tenderers};
-pub use crate::indicators::{Calculate, Codelist, Group, Indicator, Indicators, Settings};
+pub use crate::indicators::{Calculate, Codelist, Group, Indicator, Indicators, Modifications, Settings};
 use crate::queue::Job;
 use crate::standard::{AWARD_STATUS, BID_STATUS};
 
@@ -472,6 +472,16 @@ impl Prepare {
         let corrections = settings.corrections.unwrap_or_default();
         let award_status_by_contract_status = corrections.award_status_by_contract_status.unwrap_or_default();
 
+        // [modifications.split]
+        let modifications = settings.modifications.unwrap_or_default();
+        let mut procurement_method_details_pat = None;
+        match modifications {
+            Modifications::None => {}
+            Modifications::Split {
+                procurement_method_details,
+            } => procurement_method_details_pat = procurement_method_details,
+        };
+
         // [codelists.*]
         let codelists = settings.codelists.unwrap_or_default();
         let bid_status = codelists.get(&Codelist::BidStatus).unwrap_or(&default_mapping);
@@ -509,6 +519,12 @@ impl Prepare {
             // /tender
             if let Some(Value::Object(tender)) = release.get_mut("tender") {
                 prepare_id_object!(tender, "procuringEntity", redact_organization_id);
+
+                if let Some(pat) = &procurement_method_details_pat
+                    && let Some(Value::String(procurement_method_details)) = tender.get_mut("procurementMethodDetails")
+                {
+                    *procurement_method_details = procurement_method_details.split(pat).next().unwrap().trim_end().into();
+                }
             }
 
             // /bids